+++ src/signals.f90
@@ -8,7 +8,7 @@ module signals
 
     private
 
-    public :: add_sinusoidal_signal
+    public :: add_sinusoidal_signal, add_karplus_strong
 
 contains
 
@@ -42,13 +42,38 @@ contains
         end do
     end subroutine add_sinusoidal_signal
 
+
+    subroutine add_karplus_strong(track, t1, t2, f, Amp)
+      ! Karplus and Strong algorithm (1983), for plucked-string
+      ! http://crypto.stanford.edu/~blynn/sound/karplusstrong.html
+      ! https://en.wikipedia.org/wiki/Karplus%E2%80%93Strong_string_synthesis
+      integer, intent(in) :: track
+      real(kind=dp), intent(in) :: t1, t2, f, Amp
+      integer :: i, P
+      real(kind=dp) :: signal, r
+
+      P = int(RATE / f) - 2
+
+      ! Initial noise:
+      do i = int(t1*RATE), int(t1*RATE) + P
+          call random_number(r)
+          signal = Amp * (2*r - 1.0_dp)
+          left(track, i)  = signal
+          right(track, i) = signal
+      end do
+      ! Delay and decay:
+      do i= int(t1*RATE) + P + 1, int(t2*RATE) - 1
+          left(track, i)  = Amp/2 * (left(track, i-P) + left(track, i-P-1))
+          right(track, i) = left(track, i)
+      end do
+  end subroutine add_karplus_strong
+
     !void add_square_wave(int track, double t1, double t2, double f, double Amp) {
     !void add_sawtooth_wave(int track, double t1, double t2, double f, double Amp) {
     !void add_reverse_sawtooth_wave(int track, double t1, double t2, double f, double Amp)
     !void add_triangle_wave(int track, double t1, double t2, double f, double Amp) {
     !void add_noise(int track, double t1, double t2, double Amp) {
     !void add_weird_signal(int track, double t1, double t2, double f, double Amp, unsigned int modulation) {
-    !void add_karplus_strong(int track, double t1, double t2, double f, double Amp) {
     !void add_karplus_strong_drum(int track, double t1, double t2, double f, double Amp) {
     !void add_percussion(int track, double t1, double t2, double f, double Amp, unsigned int numero) {
     !void add_weierstrass_signal(int track, double t1, double t2, double f, double Amp) {